transaction and tag page init
