+++ lib/features/history/ui/widgets/bottomSheets/bottomsheet_tags.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-
-import '../../../../../theme/theme.dart';
 import '../../../../../widgets/bottomsheet/bottomsheet_scaffold.dart';
 import '../../pages/filters/tag_page.dart';
 
@@ -11,8 +9,6 @@ Future showTagsBottomSheet(BuildContext context) async {
   const allSelected = true;
   await showScaffoldBottomsheet(
     context,
-    title: 'All Tags',
-    actionButton: const Icon(MinyIcons.plus),
     child: TagPage(
       theme: theme,
       allSelected: allSelected,
